test: add CreatePostController unit tests

+++ src/application/controllers/post/delete.ts
@@ -1,31 +1,66 @@
+import { Forbidden, NotFound } from '@/application/errors'
 import { Controller, noContent } from '@/application/protocols'
+import type { AuthenticatedRequest } from '@/application/protocols/http/authenticated-request'
 import type { HttpResponse } from '@/application/protocols/http/responses'
-import type { Post } from '@/data/usecases'
-import { ValidationBuilder as builder, type Validator } from '../../validation'
+import {
+  ValidationBuilder as builder,
+  type Validator
+} from '@/application/validation'
+import type { Post, User } from '@/data/usecases'
+import { UserModel } from '@/domain/models'
 
-type PostManager = Post.DeletePost
+type PostManager = Post.DeletePost & Post.FindPost
+type UserManager = User.FindUserByIdOrFail
+type PerformParams = AuthenticatedRequest<Post.DeleteParams>
 
 export class DeletePostController extends Controller {
-  constructor(private readonly postManager: PostManager) {
+  constructor(
+    private readonly postManager: PostManager,
+    private readonly userManager: UserManager
+  ) {
     super()
   }
 
   async perform({
+    userId,
     id
-  }: Post.DeleteParams): Promise<HttpResponse<Post.CreateResult>> {
-    await this.postManager.deletePost({
+  }: PerformParams): Promise<HttpResponse<Post.CreateResult>> {
+    const post = await this.postManager.findPost({
       id
     })
 
-    return noContent()
+    if (!post) {
+      throw new NotFound('Post not found')
+    }
+
+    const user = await this.userManager.findUserByIdOrFail(userId)
+    const isUserAllowedToDeletePost = post.user.id === user.id
+    const isUserAdmin = user.level === UserModel.Level.ADMIN
+
+    if (isUserAdmin || isUserAllowedToDeletePost) {
+      await this.postManager.deletePost({
+        id
+      })
+
+      return noContent()
+    }
+
+    throw new Forbidden('You are not allowed to delete this post')
   }
 
-  override buildValidators({ id }: Post.DeleteParams): Validator[] {
+  override buildValidators({ id, userId }: PerformParams): Validator[] {
     return [
       ...builder
         .of({
           value: id,
           fieldName: 'id'
+        })
+        .required()
+        .build(),
+      ...builder
+        .of({
+          value: userId,
+          fieldName: 'userId'
         })
         .required()
         .build()
